+++ src/Utilities/Reflection.php
@@ -662,7 +662,7 @@ class Reflection
     }
 
     /**
-     * Sets value of given property
+     * Sets value of given property in given object
      *
      * @param mixed  $object   Object that should contains given property
      * @param string $property Name of the property
@@ -690,6 +690,27 @@ class Reflection
 
         if ($notAccessible) {
             $reflectionProperty->setAccessible(false);
+        }
+    }
+
+    /**
+     * Sets values of properties in given object
+     *
+     * @param mixed $object           Object that should contains given property
+     * @param array $propertiesValues Key-value pairs, where key - name of the property, value - value of the property
+     */
+    public static function setPropertiesValues($object, array $propertiesValues)
+    {
+        /*
+         * No properties?
+         * Nothing to do
+         */
+        if (empty($propertiesValues)) {
+            return;
+        }
+
+        foreach ($propertiesValues as $property => $value) {
+            static::setPropertyValue($object, $property, $value);
         }
     }
 }